backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.86.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.42.0",
]
